requirements.txt
django==2.2.10
djangorestframework==3.12.2
pyyaml==5.3.1
uritemplate==3.0.1